# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Geevheit Intelligence 360 API
  version: 0.1.0
  description: API-first contract for the enterprise transportation intelligence platform.
paths:
...
  /api/v1/checklists:
    get:
      summary: Search checklists
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Checklist page
    post:
      summary: Create checklist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Checklist"
      responses:
        "201":
          description: Checklist created
  /api/v1/checklists/{id}:
    get:
      summary: Get checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist
    put:
      summary: Update checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist updated
    delete:
      summary: Soft delete checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Checklist deleted
  /api/v1/checklists/{id}/start:
    post:
      summary: Start checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist started
  /api/v1/checklists/{id}/finish:
    post:
      summary: Finish checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist finished
  /api/v1/checklists/{id}/cancel:
    post:
      summary: Cancel checklist
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist cancelled
  /api/v1/checklists/{id}/items:
    get:
      summary: List checklist items
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist items
    post:
      summary: Add checklist item
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistItem"
      responses:
        "201":
          description: Checklist item created
  /api/v1/checklists/{id}/items/{item_id}:
    put:
      summary: Update checklist item
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/ItemID"
      responses:
        "200":
          description: Checklist item updated
    delete:
      summary: Delete checklist item
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/ItemID"
      responses:
        "204":
          description: Checklist item deleted
  /api/v1/checklists/{id}/answers:
    get:
      summary: List checklist answers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist answers
    post:
      summary: Answer checklist item
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistAnswer"
      responses:
        "201":
          description: Checklist answer created
...
  /api/v1/checklists/templates:
    get:
      summary: Search checklist templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist template page
    post:
      summary: Create checklist template
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Checklist template created
  /api/v1/checklists/templates/{id}/versions:
    get:
      summary: List checklist template versions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Template version page
    post:
      summary: Create checklist template version
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Template version created
  /api/v1/checklists/templates/versions/{id}/publish:
    post:
      summary: Publish checklist template version
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Template version published
  /api/v1/checklists/types:
    get:
      summary: Search checklist types
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist type page
    post:
      summary: Create checklist type
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Checklist type created
  /api/v1/checklists/sections:
    get:
      summary: Search checklist sections
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist section page
    post:
      summary: Create checklist section
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Checklist section created
  /api/v1/checklists/items:
    get:
      summary: Search checklist engine items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist engine item page
    post:
      summary: Create checklist engine item
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Checklist engine item created
  /api/v1/checklists/executions:
    get:
      summary: Search checklist executions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checklist execution page
    post:
      summary: Start checklist execution
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Checklist execution started
  /api/v1/checklists/executions/{id}/responses:
    get:
      summary: List checklist execution responses
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist response page
    post:
      summary: Record checklist response
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Checklist response recorded
  /api/v1/checklists/executions/{id}/evidence:
    get:
      summary: List checklist execution evidence
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist evidence page
    post:
      summary: Add checklist evidence
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Checklist evidence added
  /api/v1/checklists/executions/{id}/non-conformities:
    get:
      summary: List checklist non-conformities
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist non-conformity page
    post:
      summary: Create checklist non-conformity
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Checklist non-conformity created
  /api/v1/checklists/executions/{id}/history:
    get:
      summary: List checklist execution history
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Checklist history page
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
    ItemID:
      name: item_id
      in: path
      required: true
      schema:
        type: string
...
    Search:
      name: search
      in: query
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
  schemas:
...
    Checklist:
      type: object
      required: [vehicle_id, title]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        vehicle_id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
    ChecklistItem:
      type: object
      required: [title]
      properties:
        id:
          type: string
        checklist_id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        required:
          type: boolean
        order_index:
          type: integer
        answer_type:
          type: string
        expected_value:
          type: string
    ChecklistAnswer:
      type: object
      required: [checklist_item_id]
      properties:
        id:
          type: string
        checklist_item_id:
          type: string
        answer:
          type: string
        notes:
          type: string
        photo_url:
          type: string
...